Fixes #19960: Remove useless bootchecks in rudder 7.0
